+++ src/setting.js
@@ -3,5 +3,7 @@ module.exports = {
     // custom request function
     request: void 0,
     // custom extend function
-    extend: void 0
+    extend: void 0,
+    // whether current mode is debug
+    debug: !0
 };